feat(caja): sobreventa sin tope + Por conciliar + print stock bajo
- venta: quita tope de stock en front (useCarrito/tabla) y backend
  (new_venta sin guard stock>=?); el stock puede quedar negativo
- modelos: CartItemRequest.id i32->i64 (alineado al rowid de SQLite)
- tickets: get_mi_ticket_detalle tolera venta_id/ventaId + guard en front
- inventario: estado Por conciliar si sistema<0 (pantalla + ticket),
  fisico inicial saneado a >=0, print ya no aplana el negativo
- impresora: comando imprimir_lista_stock_bajo + boton en la Alerta
- docs: seccion 11 conciliacion en ComoFunciona.md (facilito)
- tests: ventas/estres actualizados a sobreventa; tsc+vitest+cargo en verde

+++ yarvis-app/src-tauri/src/models.rs
@@ -87,7 +87,7 @@ pub struct EmployeeGoal {
 
 #[derive(Deserialize, Clone)]
 pub struct CartItemRequest {
-    pub id: Option<i32>,
+    pub id: Option<i64>,
     pub nombre: String,
     pub precio_venta: f64,
     pub cantidad: f64,
